monomorphise-skeleton: use the same scoping environment for all conditions in the same value
When localising the condition for, e.g., a return value, we can re-use
the same existentials for all references to the same binding. Of course,
we don't want to let that sharing cross the elaboration of potential values
via dataflow.

This lets us remember that, e.g., (x, y) -> (- x y), where (= y x)
always has a result of zero.